% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_option}
\alias{add_option}
\title{Add option to the model}
\usage{
add_option(vwmodel, option = c("binary", "oaa", "ect", "csoaa", "wap",
  "log_multi", "lda", "mf", "lrq", "stage_poly", "bootstrap", "autolink", "cb",
  "cbify", "nn", "topk", "struct_search", "boosting", "ksvm"), ...)
}
\arguments{
\item{vwmodel}{Model of vw class}

\item{option}{Name of an option}

\item{...}{Additional options for a learning algorithm / reduction}
}
\description{
Add a learning algorithm / reduction to the option stack inside model
}
