% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_reduction}
\alias{add_reduction}
\title{Add reduction to the model}
\usage{
add_reduction(vwmodel, reduction = c("binary", "oaa", "ect", "csoaa", "wap",
  "log_multi", "lda", "mf", "lrq", "stage_poly", "bootstrap", "autolink", "cb",
  "cbify", "nn", "topk", "struct_search", "boosting", "ksvm"), ...)
}
\arguments{
\item{vwmodel}{Model of vw class}

\item{reduction}{Name of reduction}

\item{...}{Reduction options}
}
\description{
Add reduction to the reduction stack inside model
}
