% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fromVw}
\alias{fromVw}
\title{Convert Vowpwal Wabbit Dataset to List or Data Frame}
\usage{
fromVw(x, dense = FALSE, keepvw = FALSE)
}
\arguments{
\item{x}{vector of Vowpal Wabbit strings.}

\item{dense}{whether to return the data in dense format as a data frame, or in a sparse
format as a list. Defaults to false.}

\item{keepvw}{whether to include the original Vowpal Wabbit string in the output.
Defaults to false.}
}
\value{
Data frame of examples including columns label, importance, tag, feature
  values, and optionally the raw input string.
}
\description{
Convert a vector of strings in Vowpal Wabbit format to a list of sparse data frames or
to a dense data frame.
}
\details{
Vowpal Wabbit strings are in the format:

[Label] [Importance [Tag]]|Namespace Features |Namespace Features ... |Namespace Features

where

\itemize{
  \item Label is an real number representing the dependent variable. Because examples can be unlabeled, Label is optional.
  \item Importance is the non-negative importance weight, and defaults to 1.0 if unspecified.
  \item Tag is an optional tag for the example, and defaults to NA if unspecified.
  \item Namespace=String[:Value] is a namespace string optionally followed by a float representing the global weight for features in the namespace. If the weight is omitted, 1.0 is substituted.
  \item Features=(String[:Value] )* is a sequence of feature names optionally followed by a float representing the feature weight. If the weight is omitted, 1.0 is substituted.
}

Conversion to R objects occurs by concatenating the namespace string with each feature
name in that namespace, and multiplying the namespace global weight with the feature
weight.

Additional details on the Vowpal Wabbit format can be found here:
\url{https://github.com/JohnLangford/vowpal_wabbit/wiki/Input-format}.
}
\examples{
vw <- c(
  "1 2 'tag|a:2 b:3",
  "0 |f:.23 sqft:.25 age:.05 2006"
)

fromVw(x = vw, dense = FALSE)
# [[1]]
#   label importance  tag a_b
# 1     1          2 'tag   6
#
# [[2]]
#   label importance  tag f_sqft  f_age f_2006
# 1     0          1 <NA> 0.0575 0.0115   0.23

fromVw(x = vw, dense = TRUE)
#   label importance  tag a_b f_sqft  f_age f_2006
# 1     1          2 'tag   6     NA     NA     NA
# 2     0          1 <NA>  NA 0.0575 0.0115   0.23
}

