% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot a Density Function for the Given Prediction and Actual}
\usage{
plotDensity(x)
}
\arguments{
\item{x}{A data.frame object containing (at least) two columns
\sQuote{predicted} and \sQuote{actual}}
}
\value{
A ggplot2 object is returned, but the function is invoked
 for the side effect of displaying the plot
}
\description{
Simple density plotting function for classification outcome
}
\references{
See Zumel and Mount, Practical Data Science with R,
 page 101
}
\author{
Dirk Eddelbuettel
}
