% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.vw}
\alias{print.vw}
\title{Print VW model}
\usage{
\method{print}{vw}(x, ...)
}
\arguments{
\item{x}{Model of vw class}

\item{...}{Not used currently}
}
\description{
Print information about Vowpal Wabbit model
}
\examples{
vwmodel <- vwsetup()
print(vwmodel)

}
