% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rvw}
\alias{rvw}
\title{Convert Vowpal Wabbit datasets to R objects.}
\description{
rvw converts data in Vowpwal Wabbit format to R lists of sparse data frames or to dense
data frames. This is useful in conjunction with downsampling big data in Vowpal Wabbit
format for prototyping models in R and conducting exploratory data anlaysis. A to-do
item is to add conversion from R data frames to Vowpwal Wabbit format, which should be
straightforward. Bug reports and contributions welcome.
}
\details{
Additional details on the Vowpal Wabbit format can be found here:
\url{https://github.com/JohnLangford/vowpal_wabbit/wiki/Input-format}.
}

