\name{rvwgsoc-package}
\alias{rvwgsoc-package}
\alias{rvwgsoc}
\docType{package}
\title{
  R interface for Vowpal Wabbit
}
\description{
  R interface for Vowpal Wabbit using Rcpp and libvw for GSoC 2018.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Ivan Pavlov, Dirk Eddelbuettel, James J Balamuta

Maintainer: Ivan Pavlov <pavlov.ivan095@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
