% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vwNamespacesToList}
\alias{vwNamespacesToList}
\title{Convert the Vowpal Wabbit 'Namespace Features |Namespace Features ... |Namespace
Features' to a list}
\usage{
vwNamespacesToList(x)
}
\arguments{
\item{x}{Namespace Features |Namespace Features ... |Namespace Features}
}
\value{
list of lists with names equal to the namespace prepended to feature names and
  values equal to the product of the namespace weight and feature weights.
}
\description{
Convert the Vowpal Wabbit 'Namespace Features |Namespace Features ... |Namespace
Features' to a list
}
\examples{
# Returns list(namespace1_feature1 = 4.0, namespace1_feature2 = 6.0,
  # namespace2_feature1 = 1.0, namespace2_feature3 = 1.0)
  vwNamespacesToList('namespace1:2.0 feature1:2.0 feature2:3.0 |namespace2 feature1 feature3')
}
\keyword{internal}

