% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vwNamespaceToList}
\alias{vwNamespaceToList}
\title{Convert Vowpal Wabbit 'Namespace Features ' section to a list}
\usage{
vwNamespaceToList(x)
}
\arguments{
\item{x}{Vowpal Wabbit 'Namespace Features ' section where Namespace=String[:Value] and
Features=(String[:Value] )*. If a weight is omitted, 1.0 is substituted.}
}
\value{
list with names equal to the namespace prepended to feature names and values
  equal to the product of the namespace weight and feature weights.
}
\description{
Converts Vowpal Wabbit 'Namespace Features ' section by concatenating the namespace
string with each feature name and multiplying the namespace global weight with the
feature weight. Namespace=String[:Value] is a namespace string optionally followed by a
float representing the global weight for features in the namespace, and
Features=(String[:Value] )* is a sequence of feature names optionally followed by a
float representing the feature weight. If a weight is omitted, 1.0 is substituted.
}
\examples{
# Returns list(namespace1_feature1 = 4.0, namespace1_feature2 = 6.0)
  vwNamespaceToList('namespace1:2.0 feature1:2.0 feature2:3.0 ')
  # Returns list(namespace2_feature1 = 1.0, namespace2_feature3 = 1.0)
  vwNamespaceToList('namespace2 feature1 feature3 ')
}
\keyword{internal}

