% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vwparams}
\alias{vwparams}
\alias{vwparams<-}
\title{Access and modify parameters of VW model}
\usage{
vwparams(vwmodel, name)

vwparams(vwmodel, name) <- value
}
\arguments{
\item{vwmodel}{Model of vw class}

\item{name}{Name of VW parameter}

\item{value}{Replacment value of a parameter}
}
\value{
Value of a parameter
}
\description{
These functions allow to access VW model parameters by name and correctly modify them
}
\examples{
vwmodel <- vwsetup()
# Access parameter
vwparams(vwmodel, "bit_precision")
# Modify parameter
vwparams(vwmodel, "bit_precision") <- 25

}
