% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vwtest}
\alias{vwtest}
\title{Compute predictions using Vowpal Wabbit model}
\usage{
vwtest(vwmodel, data_path = "", probs_path = "")
}
\arguments{
\item{vwmodel}{Model of vw class}

\item{data_path}{Path to validation data in .vw plain text format}

\item{probs_path}{Path to file where to save predictions}
}
\value{
Numerical vector containing predictions
}
\description{
Compute predictions using Vowpal Wabbit model
}
\examples{
# vwmodel <- vwsetup(dir="./tests/testthat/")
# vwpredicted <- vwtest(vwmodel, "X_test.vw")
}
