% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vwtrain}
\alias{vwtrain}
\title{Train Vowpal Wabbit model}
\usage{
vwtrain(vwmodel, data_path = "")
}
\arguments{
\item{vwmodel}{Model of vw class to train}

\item{data_path}{Path to training data in .vw plain text format.}
}
\description{
Train Vowpal Wabbit model
}
\examples{
# vwmodel <- vwsetup(dir="./tests/testthat/")
# vwtrain(vwmodel, "X_train.vw")
}
