% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df2vw}
\alias{df2vw}
\title{Create a VW data file from a R data.frame object}
\usage{
df2vw(data, file_path, namespaces = NULL, keep_space = NULL,
  fixed = NULL, targets = NULL, probabilities = NULL,
  weight = NULL, base = NULL, tag = NULL, multiline = NULL,
  append = FALSE)
}
\arguments{
\item{data}{[data.frame] data.frame object to be converted}

\item{file_path}{[string] file name of the resulting data in
VW-friendly format}

\item{namespaces}{[list or yaml file] name of each namespace and
each variable for each namespace can be a R list, or a YAML
file example namespace with the IRIS database: namespaces =
list(sepal = list('Sepal.Length', 'Sepal.Width'), petal = list('Petal.Length',
'Petal.Width') this creates 2 namespaces (sepal
and petal) containing the features defined by elements of this lists.}

\item{keep_space}{[string vector] keep spaces for this features
Example:"FERRARI 4Si"
With \code{keep_space} will be "FERRARI 4Si" and will be treated as two features
Without \code{keep_space} will be "FERRARI_4Si" and will be treated as one feature}

\item{fixed}{[string vector] fixed parsing for this features
Similar to \code{keep_space}, but parse features exactly without replacement of special characters ("(", ")", "|", ":", "'").
Can be used for LDA ("word_1:2 word_2:3" will stay the same),
but should be used carefully, because special characters can ruin final VW format file.}

\item{targets}{[string or string vector]
If \code{[string]} then will be treated as vector with real number labels for regular VW input format.
If \code{[string vector]} then will be treated as vectors with class costs for wap and csoaa
multi-class classification algorithms or as vectors with actions for Contextual Bandit algorithm.}

\item{probabilities}{[string vector] vectors with action probabilities for Contextual Bandit algorithm.}

\item{weight}{[string] weight (importance) of each line of the dataset.}

\item{base}{[string] base of each line of the dataset. Used for residual regression.}

\item{tag}{[string] tag of each line of the dataset.}

\item{multiline}{[integer] number of labels (separate lines) for multilines example}

\item{append}{[bool] data to be appended to the result file}
}
\description{
Create a VW data file from a R data.frame object
}
