% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vwtest}
\alias{vwtest}
\title{Compute predictions using Vowpal Wabbit model}
\usage{
vwtest(vwmodel, data_path = "", probs_path = "", readable_model = NULL,
  quiet = FALSE)
}
\arguments{
\item{vwmodel}{Model of vw class to train}

\item{data_path}{Path to training data in .vw plain text format}

\item{probs_path}{Path to file where to save predictions}

\item{readable_model}{Print trained model in human readable format ("hashed") 
and also with human readable features ("inverted")}

\item{quiet}{Do not print anything to the console}
}
\value{
Numerical vector containing predictions
}
\description{
Compute predictions using Vowpal Wabbit model
}
\examples{
ext_train_data <- system.file("extdata", "X_train.vw", package = "rvwgsoc")
ext_test_data <- system.file("extdata", "X_valid.vw", package = "rvwgsoc") 
test_vwmodel <- vwsetup()
vwtrain(test_vwmodel, data_path = ext_train_data)
vwtrain(test_vwmodel, data_path = ext_test_data)
}
