% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/functions.R
\name{vwtest}
\alias{vwtest}
\alias{predict.vw}
\title{Compute predictions using Vowpal Wabbit model}
\usage{
vwtest(vwmodel, data, probs_path = "", full_probs = FALSE,
  readable_model = NULL, readable_model_path = "", quiet = FALSE,
  passes = 1L, cache = FALSE, raw = FALSE, progress = NULL,
  namespaces = NULL, keep_space = NULL, fixed = NULL,
  targets = NULL, probabilities = NULL, weight = NULL, base = NULL,
  tag = NULL, multiline = NULL)

\method{predict}{vw}(object, data, probs_path = "", full_probs = FALSE,
  readable_model = NULL, quiet = FALSE, ...)
}
\arguments{
\item{vwmodel}{Model of vw class to train.}

\item{data}{[string or data.frame] Path to training data in .vw plain text format or data.frame.
If \code{[data.frame]} then will be parsed using \code{df2vw} function.}

\item{probs_path}{[string] Path to file where to save predictions.}

\item{full_probs}{[bool] Output full predictions in data.frame format. If not, force predictions into a single vector (default).}

\item{readable_model}{[string] Print trained model in human readable format ("hashed") 
and also with human readable features ("inverted").}

\item{readable_model_path}{[string] Path to file where to save readable model.}

\item{quiet}{[bool] Do not print anything to the console.}

\item{passes}{[int] Number of times the algorithm will cycle over the data (epochs).}

\item{cache}{[bool] Use a cache for a data file.}

\item{raw}{[bool] Output unnormalized predictions. Default is FALSE.}

\item{progress}{[integer/real] Progress update frequency. int: additive, real: multiplicative}

\item{namespaces}{[list or yaml file] For \code{df2vw}. Name of each namespace and
each variable for each namespace can be a R list, or a YAML
file example namespace with the IRIS database: namespaces =
list(sepal = list('Sepal.Length', 'Sepal.Width'), petal = list('Petal.Length',
'Petal.Width') this creates 2 namespaces (sepal
and petal) containing the features defined by elements of this lists.}

\item{keep_space}{[string vector] For \code{df2vw}. Keep spaces for this features
Example:"FERRARI 4Si"
With \code{keep_space} will be "FERRARI 4Si" and will be treated as two features
Without \code{keep_space} will be "FERRARI_4Si" and will be treated as one feature}

\item{fixed}{[string vector] fixed parsing for this features
Similar to \code{keep_space}, but parse features exactly without replacement of special characters ("(", ")", "|", ":", "'").
Can be used for LDA ("word_1:2 word_2:3" will stay the same),
but should be used carefully, because special characters can ruin final VW format file.}

\item{targets}{[string or string vector] For \code{df2vw}.
If \code{[string]} then will be treated as vector with real number labels for regular VW input format. 
If \code{[string vector]} then will be treated as vectors with class costs for wap and csoaa 
multi-class classification algorithms or as vectors with actions for Contextual Bandit algorithm.}

\item{probabilities}{[string vector] For \code{df2vw}. Vectors with action probabilities for Contextual Bandit algorithm.}

\item{weight}{[string] For \code{df2vw}. Weight (importance) of each line of the dataset.}

\item{base}{[string] For \code{df2vw}. Base of each line of the dataset. Used for residual regression.}

\item{tag}{[string] For \code{df2vw}. Tag of each line of the dataset.}

\item{multiline}{[integer] Number of labels (separate lines) for multilines example}

\item{object}{Model of vw class to train for \code{predict.vw}}

\item{...}{Parameters passed to \code{predict.vw}}
}
\value{
Numerical vector containing predictions
}
\description{
\code{vwtest} computes predictions using VW model from \code{\link{vwsetup}}
\code{predict.vw} compute predictions using parser settings from \code{\link{vwtrain}}
}
\examples{
ext_train_data <- system.file("extdata", "binary_train.vw", package = "rvw")
ext_test_data <- system.file("extdata", "binary_valid.vw", package = "rvw") 
test_vwmodel <- vwsetup()
vwtrain(test_vwmodel, data = ext_train_data)
vwtest(test_vwmodel, data = ext_test_data)
}
