% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vwtrain}
\alias{vwtrain}
\title{Train Vowpal Wabbit model}
\usage{
vwtrain(vwmodel, data_path = "", readable_model = NULL, quiet = FALSE)
}
\arguments{
\item{vwmodel}{Model of vw class to train}

\item{data_path}{Path to training data in .vw plain text format}

\item{readable_model}{Print trained model in human readable format ("hashed") 
and also with human readable features ("inverted")}

\item{quiet}{Do not print anything to the console}
}
\description{
Train Vowpal Wabbit model
}
\examples{
ext_train_data <- system.file("extdata", "X_train.vw", package = "rvwgsoc")
test_vwmodel <- vwsetup()
vwtrain(test_vwmodel, data_path = ext_train_data)
}
